Move Heya models to new app
